GRAD-25
This commit is the activity diagram for the survey that users will take and then the data will be stored in an excel spreadsheet. It will be formatted as a user friendly excel sheet that can help users budget based on income, expenses, debt, and investments. This then will be a excel sheet users can use to budget in their day to day lives.